+++ Paardensprong-SFML/CellGrid.h
@@ -6,9 +6,9 @@
 
 
 struct Cell {
-    float x{ 0.0 }, y{ 0.0 };
-    float size{ 96.0 };
-    float border_width = size / 12;
+    f32 x{ 0.0 }, y{ 0.0 };
+    f32 size{ 96.0 };
+    f32 border_width = size / 12;
 
     sf::Color bg_color{ sf::Color::Red };
     sf::Color text_color{ sf::Color::White };
@@ -20,9 +20,9 @@ struct Cell {
 
 
     void paint(sf::RenderWindow& window) {
-        float real_size = size - border_width * 2;
-        float real_x = x + border_width;
-        float real_y = y + border_width;
+        f32 real_size = size - border_width * 2;
+        f32 real_x = x + border_width;
+        f32 real_y = y + border_width;
         
         {
             sf::RectangleShape shape = sf::RectangleShape({ real_size, real_size });
@@ -32,7 +32,7 @@ struct Cell {
         }
 
         {
-            float text_size = real_size - 16;
+            u16 text_size = (u16)real_size - 16;
             sf::Text text_shape = sf::Text(letter, font, text_size);
             text_shape.setStyle(sf::Text::Bold);
 
@@ -42,8 +42,8 @@ struct Cell {
                 bounds.top  + bounds.height/2.0f
             );
 
-            float center_x = real_x + (real_size / 2.0f);
-            float center_y = real_y + (real_size / 2.0f);
+            f32 center_x = real_x + (real_size / 2.0f);
+            f32 center_y = real_y + (real_size / 2.0f);
 
             text_shape.setFillColor(text_color);
             text_shape.setPosition(center_x, center_y);
@@ -53,27 +53,27 @@ struct Cell {
 };
 
 struct CellGrid {
-    static constexpr int GRID_COUNT = 9;
+    static constexpr u32 GRID_COUNT = 9;
     std::vector<Cell> grid = std::vector<Cell>(GRID_COUNT);
 
 
-    float x{ 0.0 }, y{ 0.0 };
-    float dimension{ 3.0 };
-    float cell_size{ 96.0 };
-    float size{ cell_size * dimension };
+    f32 x{ 0.0 }, y{ 0.0 };
+    f32 dimension{ 3.0 };
+    f32 cell_size{ 96.0 };
+    f32 size{ cell_size * dimension };
 
-    int world_dim_w = 800;
-    int world_sim_h = 600;
+    u32 world_dim_w = 800;
+    u32 world_sim_h = 600;
 
     CellGrid(std::string letters[]) {
         setupGrid(letters);
     }
 
-    int getGridIndexForLetterIndex(int letter_index);
+    u16 getGridIndexForLetterIndex(u16 letter_index);
 
     void setupGrid(std::string letters[]);
 
-    void position(float new_x, float new_y);
+    void position(f32 new_x, f32 new_y);
 
     void paint(sf::RenderWindow& window);
 };
